requirements.txt
gensim
mecab-python3
sklearn
